api/grant-query/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-google-alloydb-pg
google-cloud-alloydb-connector[pg8000]
pg8000
sqlalchemy
sqlmodel
